Changes for demos
Change target_publisher output topic name from '/point_command' to
'/goal_object_position' as it is more descriptive. This script is used
for the following demo.

Remove calibration_testing script. It was previously finalized and
renamed to 'calibration_publisher'.

Add 'target_publisher_usb' script, which is used for the servo demos.